requirements.txt
absl-py==2.0.0
appnope==0.1.4
asgiref==3.5.2
asttokens==2.4.1
astunparse==1.6.3
attrs==23.1.0
beautifulsoup4==4.12.2
Brotli==1.0.9
bs4==0.0.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
contourpy==1.2.0
cryptography==38.0.4
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
Flask==2.2.5
flatbuffers==23.5.26
fonttools==4.47.0
fsspec==2024.2.0
gast==0.5.4
google-auth==2.26.1
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.0
h11==0.14.0
h2==4.1.0
h5py==3.10.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipykernel==6.29.2
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
kaitaistruct==0.10
keras==2.15.0
kiwisolver==1.4.5
ldap3==2.9.1
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mitmproxy==9.0.1
mitmproxy_wireguard==0.1.23
ml-dtypes==0.2.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.0.8
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.2
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.4
parso==0.8.3
passlib==1.7.4
pexpect==4.9.0
Pillow==10.1.0
pip==23.3.2
platformdirs==4.2.0
pproxy==2.7.9
prompt-toolkit==3.0.43
protobuf==4.23.4
psutil==5.9.8
ptyprocess==0.7.0
publicsuffix2==2.20191221
pure-eval==0.2.2
pvpn==0.2.5
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.21
pycryptodome==3.20.0
PyGetWindow==0.0.9
Pygments==2.17.2
PyMsgBox==1.0.9
pyobjc-core==10.1
pyobjc-framework-Cocoa==10.1
pyobjc-framework-Quartz==10.1
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
pytweening==1.0.7
pytz==2023.3.post1
pyzmq==25.1.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
rubicon-objc==0.4.7
scikit-learn==1.3.2
scipy==1.11.4
selenium==4.16.0
selenium-stealth==1.0.6
setuptools==58.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.15.0
termcolor==2.4.0
threadpoolctl==3.2.0
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
tornado==6.4
traitlets==5.14.1
trio==0.23.2
trio-websocket==0.11.1
typing_extensions==4.4.0
tzdata==2023.3
undetected-chromedriver @ git+https://www.github.com/ultrafunkamsterdam/undetected-chromedriver@783b8393157b578e19e85b04d300fe06efeef653
urllib3==2.1.0
urwid==2.1.2
watchdog==3.0.0
wcwidth==0.2.13
webdriver-manager==4.0.1
websockets==12.0
Werkzeug==3.0.1
wheel==0.42.0
wrapt==1.14.1
wsproto==1.2.0
zipp==3.17.0
zstandard==0.19.0
